Rename and reorder for clarity
